DTSCCI-920 RefactorDirectionsQuestionnaireGenerator (#5610)
* Extracted DirectionsQuestionnaireGenerator Methods into New classes

* extract & add Tests

* Amend tests

* Add missing classes to Integration tests

* Fix assert

* Amend test for DQLipResponseGeneratorTest

* Remove Code related to hnl toggle.

* Revert "Remove Code related to hnl toggle."

This reverts commit 10ba0496bb9c93b67bf331838520c93765f87981.

* Remove TODO for hnl toogle

* Fix sonar

---------
