+++ src/test/java/uk/gov/hmcts/reform/civil/service/docmosis/dq/DirectionsQuestionnaireGeneratorTest.java
@@ -54,6 +54,10 @@ import uk.gov.hmcts.reform.civil.service.FeatureToggleService;
 import uk.gov.hmcts.reform.civil.service.docmosis.DocmosisTemplates;
 import uk.gov.hmcts.reform.civil.service.docmosis.DocumentGeneratorService;
 import uk.gov.hmcts.reform.civil.service.docmosis.RepresentativeService;
+import uk.gov.hmcts.reform.civil.service.docmosis.dq.builders.DQGeneratorFormBuilder;
+import uk.gov.hmcts.reform.civil.service.docmosis.dq.helpers.GetRespondentsForDQGenerator;
+import uk.gov.hmcts.reform.civil.service.docmosis.dq.helpers.RespondentTemplateForDQGenerator;
+import uk.gov.hmcts.reform.civil.service.docmosis.dq.helpers.SetApplicantsForDQGenerator;
 import uk.gov.hmcts.reform.civil.service.flowstate.SimpleStateFlowEngine;
 import uk.gov.hmcts.reform.civil.service.flowstate.TransitionsTestConfiguration;
 import uk.gov.hmcts.reform.civil.service.referencedata.LocationReferenceDataService;
@@ -102,7 +106,11 @@ import static uk.gov.hmcts.reform.civil.utils.ElementUtils.unwrapElements;
     SimpleStateFlowEngine.class,
     SimpleStateFlowBuilder.class,
     TransitionsTestConfiguration.class,
-    CaseDetailsConverter.class
+    CaseDetailsConverter.class,
+    SetApplicantsForDQGenerator.class,
+    GetRespondentsForDQGenerator.class,
+    RespondentTemplateForDQGenerator.class,
+    DQGeneratorFormBuilder.class
 })
 class DirectionsQuestionnaireGeneratorTest {
 
@@ -914,7 +922,6 @@ class DirectionsQuestionnaireGeneratorTest {
                     .build();
                 uk.gov.hmcts.reform.civil.model.dq.Expert expert1 =
                     uk.gov.hmcts.reform.civil.model.dq.Expert.builder()
-                        .name("Expert 1")
                         .firstName("first")
                         .lastName("last")
                         .phoneNumber("07123456789")
